Move controls to sidebar
